requirements.txt
celery==5.3.1
certifi==2023.7.22
Flask==2.3.2
Flask_Cors==4.0.0
Flask_JWT_Extended==4.5.2
Flask_PyMongo==2.3.0
Pillow==10.4.0
pymongo==4.4.1
python-dotenv==1.0.1
Requests==2.32.3
pandas
openpyxl
boto3
redis
